requirements.txt
appdirs==1.4.4
asgiref==3.3.4
boto3==1.17.76
botocore==1.20.76
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.3
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==4.0.1
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
virtualenv==20.4.6
whitenoise==5.2.0
zipp==3.4.1
